fix: start editing button did not work
Current Behavior:
When the start editing button is clicked, the button throws an error
that the URL is not valid.

Expected Behavior:
When the start editing button ic clicked, it should redirect to the CS
website.

Notes:
I had removed the dependency of `this.startEditingButton` from the
class to make the methods self sustainable and used `event.target`
instead. This returned the `<span/>` inside the button. So, I have
replaced it with `event.currentTarget` to function properly.