fix: call onChangeCallback when live_preview parameter is present in URL
Visual Builder doesn't send client-data-send on site load, this leads to
the site not making any calls to the preview service which is required
for the subsequent entry PATCH requests to work correctly.
onChangeCallback is called when either skipInitialRender is false or
when live_preview search parameter is present in the URL. Since Visual
Builder provides these values through the search parameter, the site
makes entry GET calls to the preview service. Since live preview doesn't
provide these params in non-SSR modes, it will continue to work the
same.