fix: custom space key handling when a button element is on the composed path
If there is a button element on the composed path, then the browser will send a click
event on the button when space key is pressed. stopPropagation does not work as this
is not a propagated event but something done by the browser. So, whenever a button
is detected on the event's composed path, use custom space key handling.