fix: psuedo-editable height collapse
The height of the cloned/psuedo-editable element collapses when
text-wrap-mode is set to nowrap and text-overflow is set to ellipsis.
This change ensures that we override these properties CORRECTLY.
Previously we overrode these properties but only in camel case. Now we
ensure we perform the camelCase transformation after overriding
properties. We also set min-height to whatever the current height of the
element is, to prevent height collapse in other scenarios.